♻️ refactor : enumerations 명명 choices로 변경
