+++ apps/core/choices/verification_choices.py
